feat: add composition between search components
Search is now linked with searchbar and searchdata. Data is being propagated through components.

SearchData is only a simple implementation to test links between components and needs to be fully implemented.
fetchData function in service needs to be implemented to return data that it gets, it sends good requests but doesn't return any data for now.